+++ aioftp/__main__.py
@@ -22,7 +22,7 @@ if not args["--quiet"]:
 
     logging.basicConfig(
         level=logging.INFO,
-        format="%(asctime)s %(message)s",
+        format="%(asctime)s [%(name)s] %(message)s",
         datefmt="[%H:%M:%S]:",
     )